import extracted logger
